fixed data class type issue

+++ app/src/main/java/com/example/cryptohub/model/coindata/Atl.kt
@@ -17,7 +17,7 @@ data class Atl(
     @SerializedName("bhd")
     val bhd: Double,
     @SerializedName("bits")
-    val bits: Int,
+    val bits: Double,
     @SerializedName("bmd")
     val bmd: Double,
     @SerializedName("bnb")
@@ -31,7 +31,7 @@ data class Atl(
     @SerializedName("chf")
     val chf: Double,
     @SerializedName("clp")
-    val clp: Int,
+    val clp: Double,
     @SerializedName("cny")
     val cny: Double,
     @SerializedName("czk")
@@ -51,9 +51,9 @@ data class Atl(
     @SerializedName("hkd")
     val hkd: Double,
     @SerializedName("huf")
-    val huf: Int,
+    val huf: Double,
     @SerializedName("idr")
-    val idr: Int,
+    val idr: Double,
     @SerializedName("ils")
     val ils: Double,
     @SerializedName("inr")
@@ -61,23 +61,23 @@ data class Atl(
     @SerializedName("jpy")
     val jpy: Double,
     @SerializedName("krw")
-    val krw: Int,
+    val krw: Double,
     @SerializedName("kwd")
     val kwd: Double,
     @SerializedName("link")
     val link: Double,
     @SerializedName("lkr")
-    val lkr: Int,
+    val lkr: Double,
     @SerializedName("ltc")
     val ltc: Double,
     @SerializedName("mmk")
-    val mmk: Int,
+    val mmk: Double,
     @SerializedName("mxn")
     val mxn: Double,
     @SerializedName("myr")
     val myr: Double,
     @SerializedName("ngn")
-    val ngn: Int,
+    val ngn: Double,
     @SerializedName("nok")
     val nok: Double,
     @SerializedName("nzd")
@@ -93,7 +93,7 @@ data class Atl(
     @SerializedName("sar")
     val sar: Double,
     @SerializedName("sats")
-    val sats: Int,
+    val sats: Double,
     @SerializedName("sek")
     val sek: Double,
     @SerializedName("sgd")
@@ -111,7 +111,7 @@ data class Atl(
     @SerializedName("vef")
     val vef: Double,
     @SerializedName("vnd")
-    val vnd: Int,
+    val vnd: Double,
     @SerializedName("xag")
     val xag: Double,
     @SerializedName("xau")
@@ -119,9 +119,9 @@ data class Atl(
     @SerializedName("xdr")
     val xdr: Double,
     @SerializedName("xlm")
-    val xlm: Int,
+    val xlm: Double,
     @SerializedName("xrp")
-    val xrp: Int,
+    val xrp: Double,
     @SerializedName("yfi")
     val yfi: Double,
     @SerializedName("zar")
